# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "CallCounts": {
        "properties": {
          "failed": {
            "default": 0,
            "title": "Failed",
            "type": "integer"
          },
          "in_progress": {
            "default": 0,
            "title": "In Progress",
            "type": "integer"
          },
          "success": {
            "default": 0,
            "title": "Success",
            "type": "integer"
          }
        },
        "title": "CallCounts",
        "type": "object"
      },
      "CallLabel": {
        "enum": [
          "Sales inquiry",
          "Support",
          "Complaint",
          "Appointment",
          "Follow-up",
          "Other"
        ],
        "title": "CallLabel",
        "type": "string"
      },
      "CallResponse": {
        "properties": {
          "caller_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caller Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CallLabel"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "phone_number": {
            "title": "Phone Number",
            "type": "string"
          },
          "raw_transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Transcript"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CallStatus"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "phone_number",
          "caller_name",
          "duration_seconds",
          "status",
          "summary",
          "label",
          "started_at",
          "ended_at",
          "created_at",
          "updated_at",
          "raw_transcript",
          "notes"
        ],
        "title": "CallResponse",
        "type": "object"
      },
      "CallSortField": {
        "description": "Whitelist of columns the list endpoint may sort by.\n\nClosed by construction: an unknown value is a 422 at the HTTP boundary, so the sort column can\nnever be interpolated from arbitrary input. Internal/audit columns (raw_transcript, notes, the\nTask 4 queue columns) are deliberately absent \u2014 they are neither public nor sortable.",
        "enum": [
          "created_at",
          "started_at",
          "duration_seconds",
          "caller_name",
          "phone_number",
          "status",
          "label"
        ],
        "title": "CallSortField",
        "type": "string"
      },
      "CallStatus": {
        "enum": [
          "in_progress",
          "success",
          "failed"
        ],
        "title": "CallStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PaginatedCallsResponse": {
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/CallCounts"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/CallResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "page_size",
          "total_pages",
          "counts"
        ],
        "title": "PaginatedCallsResponse",
        "type": "object"
      },
      "SortDir": {
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortDir",
        "type": "string"
      },
      "UpdateCallNotesRequest": {
        "additionalProperties": false,
        "description": "Required-but-nullable notes payload.\n\nThe ``notes`` key must be present \u2014 an empty body ``{}`` is a 422. An explicit ``null`` (or a\nblank/whitespace string, normalized in the service) clears the note. The 2000-character cap is\nenforced here, before normalization. ``extra=\"forbid\"`` makes the model fail-closed: any extra\nfield (e.g. an attempt to over-post ``status``/``summary``) is rejected with 422, so this\nendpoint can never become a mass-assignment vector.",
        "properties": {
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "notes"
        ],
        "title": "UpdateCallNotesRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebhookCallPayload": {
        "properties": {
          "call_id": {
            "format": "uuid",
            "title": "Call Id",
            "type": "string"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "raw_transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Transcript"
          },
          "status": {
            "$ref": "#/components/schemas/CallStatus"
          }
        },
        "required": [
          "call_id",
          "status"
        ],
        "title": "WebhookCallPayload",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Backend API for the Voico Calls Dashboard.",
    "title": "Voico Calls Dashboard",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/calls": {
      "get": {
        "description": "List calls. All filters are optional and combined with AND; an empty text filter is ignored.\n\n`counts` is the per-status breakdown over the active content filters (ignoring `status`), so the\nstatus tabs stay informative; `total` and the returned page additionally respect `status`.",
        "operationId": "list_calls_api_calls_get",
        "parameters": [
          {
            "description": "Exact status match.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CallStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact status match.",
              "title": "Status"
            }
          },
          {
            "description": "Case-insensitive partial match on the caller's name.",
            "in": "query",
            "name": "caller_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive partial match on the caller's name.",
              "title": "Caller Name"
            }
          },
          {
            "description": "Partial phone match. Compared on digits only, so formatting is ignored (e.g. `5552014832` matches `+1 (555) 201-4832`).",
            "in": "query",
            "name": "phone",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Partial phone match. Compared on digits only, so formatting is ignored (e.g. `5552014832` matches `+1 (555) 201-4832`).",
              "title": "Phone"
            }
          },
          {
            "description": "Exact label match (e.g. `Sales inquiry`).",
            "in": "query",
            "name": "label",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CallLabel"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact label match (e.g. `Sales inquiry`).",
              "title": "Label"
            }
          },
          {
            "description": "Minimum call duration in seconds (inclusive).",
            "in": "query",
            "name": "min_duration",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 86400,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum call duration in seconds (inclusive).",
              "title": "Min Duration"
            }
          },
          {
            "description": "Maximum call duration in seconds (inclusive).",
            "in": "query",
            "name": "max_duration",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 86400,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum call duration in seconds (inclusive).",
              "title": "Max Duration"
            }
          },
          {
            "description": "Column to sort by (whitelisted).",
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CallSortField",
              "default": "created_at",
              "description": "Column to sort by (whitelisted)."
            }
          },
          {
            "description": "Sort direction.",
            "in": "query",
            "name": "sort_dir",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortDir",
              "default": "desc",
              "description": "Sort direction."
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCallsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "Invalid filter, sort field, or duration range"
          }
        },
        "summary": "List calls with filtering, search and sorting",
        "tags": [
          "calls"
        ]
      }
    },
    "/api/calls/{call_id}": {
      "get": {
        "operationId": "get_call_api_calls__call_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "call_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Call Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Call",
        "tags": [
          "calls"
        ]
      }
    },
    "/api/calls/{call_id}/notes": {
      "patch": {
        "description": "Set or clear the free-text note on a call.\n\nSend ``{\"notes\": \"...\"}`` to set it, ``{\"notes\": null}`` (or a blank string) to clear it. The\nnote is NFC-normalized and trimmed; an absent ``notes`` key is rejected with 422.",
        "operationId": "update_call_notes_api_calls__call_id__notes_patch",
        "parameters": [
          {
            "in": "path",
            "name": "call_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Call Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCallNotesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Call not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update call notes",
        "tags": [
          "calls"
        ]
      }
    },
    "/api/webhook/call": {
      "post": {
        "description": "Update a call from its closing webhook and, for terminal calls with a transcript, enrich it.\n\nFinds the call by ``call_id`` and updates ``status``, ``duration_seconds``, ``raw_transcript``\nand ``ended_at``. When the new status is ``success``/``failed`` and a transcript is present, an\nOpenAI summary and ``CallLabel`` classification are generated and stored; if OpenAI is\nunavailable the field update still persists and ``summary``/``label`` stay null. Re-delivery is\nidempotent (no re-stamp, no re-enrichment once summarized).\n\nWhen ``WEBHOOK_SECRET`` is configured the request must carry a valid HMAC ``X-Signature`` and a\nfresh ``X-Timestamp`` (replay window); otherwise it is rejected with 401. See ADR-0007.",
        "operationId": "webhook_call_api_webhook_call_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCallPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid webhook signature (when WEBHOOK_SECRET is set)"
          },
          "404": {
            "description": "Call not found"
          },
          "422": {
            "description": "Invalid webhook payload"
          }
        },
        "summary": "Ingest a call-completion webhook (update + AI enrichment)",
        "tags": [
          "calls"
        ]
      }
    },
    "/health": {
      "get": {
        "description": "Return ``ok`` while the service is running.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Health Check Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health check",
        "tags": [
          "health"
        ]
      }
    }
  },
  "tags": [
    {
      "description": "List, inspect, annotate, and ingest call records.",
      "name": "calls"
    },
    {
      "description": "Service liveness probe.",
      "name": "health"
    }
  ]
}